Move static_getter to inferrer_utils
